Move generate_resource_copy to utility module and add basic tests
