+++ csunplugged/resources/management/commands/makeresourcethumbnails.py
@@ -10,7 +10,7 @@ from django.core.management.base import BaseCommand
 from django.template.loader import render_to_string
 from django.http.request import QueryDict
 from resources.models import Resource
-from resources.views.views import generate_resource_copy
+from resources.utils.generate_resource_copy import generate_resource_copy
 from resources.utils.get_resource_generator import get_resource_generator
 from resources.utils.resource_valid_test_configurations import resource_valid_test_configurations
 from utils.bool_to_yes_no import bool_to_yes_no
